feat: refactoring error handling for transaction verification
Ticket: WP-6189
